Add files via upload
First there is a login page and the user can login through it to home page and it is created using HTML, CSS, JAVASCRIPT and also kept validation for that form. If the user is new member there is a registration page for them and the user can register through it and it is created using HTML, CSS, JAVASCRIPT and also kept validation for that form. After that we have a home page and it is designed using HTML, CSS, JAVASCRIPT.  In the homepage we have header section in that header section we have home, courses, contact and logout page. In the courses page we have the contents about the courses we have and you can see some the topics by clicking the get started button in that page. In the contact page where any can ask about the course, fees and you can ask anything related to the course in the message box given in the contact page and it is also designed using HTML, CSS, JAVASCRIPT and validation too. After that by clicking send message you message is noted by us we will call for further updates and the page is redirected to home page again and their u can logout from this page.